chore(repo): split remaining long e2e tests
